Main GUI was created
The dimension for the gui has changed to fit the content. Store info can
now be requested from the user BUT will not be saved. The NEXT button
does not save data. Objects for store and techincians have been made.
These are probably the values that each object need. Needs to be tested
in an active enviroment to find out.